modify Claim
