+++ app/(pages)/airdrops/mock/page.tsx
@@ -14,6 +14,7 @@ import {
   formatVestingEndsAt,
   formatTemplateType,
 } from "@/app/lib/airdrop/airdropUtils";
+import Claim from "@/app/components/airdrops/Claim";
 
 export default function AirdropPage() {
   const { address, isConnecting, isReconnecting } = useRequireAccount();
@@ -50,14 +51,7 @@ export default function AirdropPage() {
       contractDeployedAt: 1731897560,
     },
   ];
-  const standardClaim = {
-    amount: BigInt(2351),
-    isClaimed: false,
-  };
-  const linearVestingClaim = {
-    claimedAmount: BigInt(1021),
-    claimable: BigInt(325),
-  };
+
   if (!isMounted || !address)
     return (
       <Center>
@@ -187,74 +181,7 @@ export default function AirdropPage() {
           </Box>
 
           {/* Status Section */}
-
-          <Box bg="#2E3748" borderRadius="md" boxShadow="md" p={4} mb={4}>
-            <VStack spacing={0.5} align="stretch">
-              <HStack justify="space-between">
-                <Text fontSize="md" fontWeight="900">
-                  {t("airdrop.yourAllocatedAmount")}
-                </Text>
-                {!isLinearVesting && !standardClaim.isClaimed && (
-                  <HStack justify="flex-start">
-                    <Text fontSize="3xl" fontWeight="medium">
-                      {`${standardClaim.amount}`}
-                    </Text>
-                    <Text fontSize="md" fontWeight="medium">
-                      YMT
-                    </Text>
-                  </HStack>
-                )}
-              </HStack>
-              {isLinearVesting && (
-                <>
-                  <HStack justify="space-between">
-                    <Text fontSize="sm">{t("airdrop.totalAmount")}</Text>
-                    <HStack justify="flex-start">
-                      <Text fontSize="3xl" fontWeight="medium">
-                        {`${linearVestingClaim.claimedAmount + linearVestingClaim.claimable}`}
-                      </Text>
-                      <Text fontSize="md" fontWeight="medium">
-                        YMT
-                      </Text>
-                    </HStack>
-                  </HStack>
-                  <HStack justify="space-between">
-                    <Text fontSize="sm">{t("dashboard.vestingDeadline")}</Text>
-                    <Text fontSize="3xl" fontWeight="medium">
-                      {currentVestingEndsAt}
-                    </Text>
-                  </HStack>
-                  <HStack justify="space-between">
-                    <Text fontSize="sm">{t("airdrop.claimed")}</Text>
-                    <HStack justify="flex-start">
-                      <Text fontSize="3xl" fontWeight="medium">
-                        {`${linearVestingClaim.claimedAmount}`}
-                      </Text>
-                      <Text fontSize="md" fontWeight="medium">
-                        YMT
-                      </Text>
-                    </HStack>
-                  </HStack>
-                  <HStack justify="space-between">
-                    <Text fontSize="sm">{t("airdrop.claimable")}</Text>
-                    <HStack justify="flex-start">
-                      <Text fontSize="3xl" fontWeight="medium">
-                        {`${linearVestingClaim.claimable}`}
-                      </Text>
-                      <Text fontSize="md" fontWeight="medium">
-                        YMT
-                      </Text>
-                    </HStack>
-                  </HStack>
-                </>
-              )}
-              {(isLinearVesting || (!isLinearVesting && !standardClaim.isClaimed)) && (
-                <Button size="sm" colorScheme="blue" width="full" mt={4}>
-                  {t("airdrop.claim")}
-                </Button>
-              )}
-            </VStack>
-          </Box>
+          <Claim isLinearVesting={isLinearVesting} currentVestingEndsAt={currentVestingEndsAt} />
 
           {/* Menu Section */}
           {isOwner && (
